- 마이페이지 UI 작업 완료 - 회원 탈퇴 UI 작업 완료 - 구매 내역 UI 작업 진행 중

+++ app/src/main/java/com/ezen/lolketing/repository/PaymentRepository.kt
@@ -131,9 +131,9 @@ class PaymentRepository @Inject constructor(
 
     suspend fun updateChargingCache(
         addCache: Long,
-        point: Int,
+        point: Long,
         grade: String,
-        accPoint: Int,
+        accPoint: Long,
         successListener: () -> Unit,
         failureListener: () -> Unit
     ) {
